Separate concerns in state/
