+++ index.html
@@ -408,7 +408,7 @@
     <script type="module" src="js/ui/modals/welcome.js"></script>
     <script type="module" src="js/ui/modals/input.js"></script>
     <script type="module" src="js/ui/modals/information.js"></script>
-    <script type="module" src="js/state/view-selector.js"></script>
+    <script type="module" src="js/ui/view-selector.js"></script>
     <script type="module" src="js/state/current-file.js"></script>
     <script type="module" src="js/ui/toggle/switch-deploy.js"></script>
   </body>
