--- js/state/load-event.js
@@ -1,116 +1,0 @@
-import { loadObjects } from "../loaders/load.js";
-import { objectTypes } from "../lib/constants/objectTypes.js";
-import { drawView } from "./view-selector.js";
-import { getViewportPosition } from "./pixi-state.js";
-import {
-  getFileData,
-  getCurrentEventIndex,
-  setCurrentEventIndex,
-  getCurrentEventNumber,
-  getCurrentEventName,
-  eventCollection,
-  currentVisObjects,
-  getEventNumbers,
-  getCurrentView,
-  saveCurrentScrollPosition,
-} from "./globals.js";
-
-function copyObject(source, destiny) {
-  for (const [key, value] of Object.entries(source)) {
-    destiny[key] = value;
-  }
-}
-function loadSelectedEvent() {
-  const currentEventIndex = getCurrentEventIndex();
-  if (eventCollection[currentEventIndex] === undefined) {
-    const objects = loadObjects(getFileData(), getCurrentEventNumber());
-
-    eventCollection[currentEventIndex] = objects;
-
-    for (const datatype in eventCollection[currentEventIndex].datatypes) {
-      const classType = objectTypes[datatype];
-      const collection =
-        eventCollection[currentEventIndex].datatypes[datatype].collection;
-      classType.setup(collection);
-    }
-    copyObject(objects, currentVisObjects);
-  } else {
-    copyObject(eventCollection[currentEventIndex], currentVisObjects);
-  }
-}
-
-export function renderEvent(eventIndex) {
-  setCurrentEventIndex(eventIndex);
-  loadSelectedEvent();
-  updateEventNumber();
-  drawView(getCurrentView());
-}
-
-export function updateEventSelectorMenu() {
-  const eventSelectorMenu = document.getElementById("event-selector-menu");
-  eventSelectorMenu.replaceChildren();
-
-  const eventNumbers = getEventNumbers();
-  for (const [eventIndex, eventNumber] of eventNumbers.entries()) {
-    const optionElementMenu = document.createElement("div");
-    optionElementMenu.className = "event-option";
-    optionElementMenu.appendChild(
-      document.createTextNode(`Event ${eventNumber}`),
-    );
-    eventSelectorMenu.appendChild(optionElementMenu);
-    optionElementMenu.addEventListener("click", () => {
-      saveCurrentScrollPosition(getViewportPosition());
-      renderEvent(eventIndex);
-      eventSelectorMenu.style.display = "none";
-    });
-  }
-}
-
-// Page updates
-const eventNumber = document.getElementById("selected-event");
-const previousEvent = document.getElementById("previous-event");
-const nextEvent = document.getElementById("next-event");
-
-function updateEventNumber() {
-  if (eventNumber.firstChild) {
-    eventNumber.removeChild(eventNumber.firstChild);
-  }
-  eventNumber.appendChild(document.createTextNode(getCurrentEventName()));
-}
-
-previousEvent.addEventListener("click", () => {
-  const eventNumbers = getEventNumbers();
-  const currentEventIndex = getCurrentEventIndex();
-
-  if (currentEventIndex <= 0) {
-    return;
-  }
-
-  const newEventNum = `${eventNumbers[currentEventIndex - 1]}`;
-  saveCurrentScrollPosition(getViewportPosition());
-  renderEvent(newEventNum);
-});
-
-nextEvent.addEventListener("click", () => {
-  const eventNumbers = getEventNumbers();
-  const currentEventIndex = getCurrentEventIndex();
-
-  if (currentEventIndex + 1 >= eventNumbers.length) {
-    return;
-  }
-
-  const newEventNum = `${eventNumbers[currentEventIndex + 1]}`;
-  saveCurrentScrollPosition(getViewportPosition());
-  renderEvent(newEventNum);
-});
-
-eventNumber.addEventListener("click", () => {
-  const eventSelectorMenu = document.getElementById("event-selector-menu");
-  if (eventSelectorMenu.style.display === "flex") {
-    eventSelectorMenu.style.display = "none";
-  } else {
-    eventSelectorMenu.style.display = "flex";
-  }
-});
-
-export { eventCollection };
